refactor(collection-seeder): prefix COVID source files with covid_, normalize RSV owned_tag
